+++ cmd/pull.go
@@ -5,6 +5,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/web-of-things-open-source/tm-catalog-cli/cmd/completion"
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/app/cli"
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/remotes"
 )
@@ -19,20 +20,24 @@ The pattern can be a full name or a prefix the names shall start with. A partial
 path parts. E.g. 'company/device' will not match 'company/device-mark2', but will match 'company/device/submodel'.
 
 Name pattern, filters and search can be combined to narrow down the result.`,
-	Args: cobra.MaximumNArgs(1),
-	Run:  executePull,
+	Args:              cobra.MaximumNArgs(1),
+	Run:               executePull,
+	ValidArgsFunction: completion.CompleteTMNames,
 }
 
 func init() {
 	RootCmd.AddCommand(pullCmd)
 	pullCmd.Flags().StringP("remote", "r", "", "name of the remote to pull from")
+	_ = pullCmd.RegisterFlagCompletionFunc("remote", completion.CompleteRemoteNames)
 	pullCmd.Flags().StringP("directory", "d", "", "TM repository directory to pull from")
+	_ = pullCmd.MarkFlagDirname("directory")
 	pullCmd.Flags().StringP("output", "o", "", "output directory where to save the pulled TMs")
+	_ = pullCmd.MarkFlagDirname("output")
+	_ = pullCmd.MarkFlagRequired("output")
 	pullCmd.Flags().StringVar(&pFilterFlags.FilterAuthor, "filter.author", "", "filter TMs by one or more comma-separated authors")
 	pullCmd.Flags().StringVar(&pFilterFlags.FilterManufacturer, "filter.manufacturer", "", "filter TMs by one or more comma-separated manufacturers")
 	pullCmd.Flags().StringVar(&pFilterFlags.FilterMpn, "filter.mpn", "", "filter TMs by one or more comma-separated mpn (manufacturer part number)")
 	pullCmd.Flags().StringVarP(&pFilterFlags.Search, "search", "s", "", "search TMs by their content matching the search term")
-	_ = pullCmd.MarkFlagRequired("output")
 }
 
 func executePull(cmd *cobra.Command, args []string) {