+++ cmd/remote/remote_remove.go
@@ -4,6 +4,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/web-of-things-open-source/tm-catalog-cli/cmd/completion"
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/app/cli"
 )
 
@@ -19,6 +20,7 @@ var remoteRemoveCmd = &cobra.Command{
 			os.Exit(1)
 		}
 	},
+	ValidArgsFunction: completion.CompleteRemoteNames,
 }
 
 func init() {