+++ internal/app/http/server.go
@@ -15,7 +15,7 @@ import (
 //       2.1. maybe reorder the properties in model.gen.go for a nicer JSON output, as oapi-codegen orders them alphabetically
 //       2.2. for path parameters "name" and "tmID", add a regex for any character -> {name:.+}, {tmID:.+}
 //       2.3. in server.gen.go, order the handler functions, in the way that the more specific routes are above the less specific
-//          e.g. r.HandleFunc(options.BaseURL+"/inventory/{name:.+}/versions" should be on top of r.HandleFunc(options.BaseURL+"/inventory/{name:.+}
+//          e.g. r.HandleFunc(options.BaseURL+"/inventory/{name:.+}/.versions" should be on top of r.HandleFunc(options.BaseURL+"/inventory/{name:.+}
 // 3. when 2. is done, comment lines "// //go:generate" again, to prevent unwanted changes by calling "go generate"
 
 // //go:generate go run github.com/deepmap/oapi-codegen/v2/cmd/oapi-codegen@v2.0.0 -package server -generate types -o server/models.gen.go ../../../api/tm-catalog.openapi.yaml