+++ cmd/list.go
@@ -16,12 +16,12 @@ var listCmd = &cobra.Command{
 	Use:   "list <NAME PATTERN>",
 	Short: "List TMs in catalog",
 	Long: `List TMs in catalog by name pattern, filters or search. 
-The pattern can be a full name or a prefix the names shall start with. A partial pattern will match only complete 
-path parts. E.g. 'company/device' will not match 'company/device-mark2', but will match 'company/device/submodel'.
+The name can be a full name or a prefix consisting of complete path parts. 
+E.g. 'MyCompany/BarTech' will not match 'MyCompany/BarTechCorp', but will match 'MyCompany/BarTech/BazLamp'.
 
 Name pattern, filters and search can be combined to narrow down the result.`,
-	Args:              cobra.MaximumNArgs(1),
-	Run:               executeList,
+	Args: cobra.MaximumNArgs(1),
+	Run:  executeList,
 	ValidArgsFunction: completion.CompleteTMNames,
 }