+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+
+
+## [v0.0.0-alpha.4]
+
 ### Added
 
 - REST API: added `meta.page.elements` to inventory response, reflecting number of entries in current result page